requirements.txt
Django>=5.0,<5.1
gunicorn
python-decouple
psycopg[binary]
psycopg2-binary
dj-database-url
requests
djangorestframework
whitenoise
pillow
tensorflow-cpu
keras
numpy
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2